+++ src/app/components/DefenceSummary.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import cn from 'classnames';
 import TranslatedComponent from './TranslatedComponent';
+import { DamageKinetic, DamageThermal, DamageExplosive } from './SvgIcons';
 
 /**
  * Defence summary
@@ -28,62 +29,41 @@ export default class DefenceSummary extends TranslatedComponent {
     let { formats, translate, units } = language;
     let hide = tooltip.bind(null, null);
 
-    let sgClassNames = cn({ muted: !ship.findInternalByGroup('sg') });
-
     return (
-      <div>
-      <h1>{translate('defence summary')}</h1>
-      <br/>
-      <table style={{ width: '100%', lineHeight: '1em', backgroundColor: 'transparent' }}>
-        <thead>
-          <tr>
-            <th rowSpan={2}>{translate('damage to')}</th>
-            <th colSpan={3}>{translate('damage from')}</th>
-          </tr>
-          <tr>
-            <th>{translate('explosive')}</th>
-            <th>{translate('kinetic')}</th>
-            <th>{translate('thermal')}</th>
-          </tr>
-        </thead>
-        <tbody>
-          { ship.shield ?
-            <tr>
-              <td className='le {sgClassNames}'>{translate('shields')}</td>
-              <td className='ri {sgClassNames}'>{ship.shieldExplRes ? formats.pct(ship.shieldExplRes) : '-'}</td>
-              <td className='ri {sgClassNames}'>{ship.shieldKinRes ? formats.pct(ship.shieldKinRes) : '-'}</td>
-              <td className='ri {sgClassNames}'>{ship.shieldThermRes ? formats.pct(ship.shieldThermRes) : '-'}</td>
-            </tr> : null }
-          <tr>
-            <td className='le'>{translate('hull')}</td>
-            <td className='ri'>{formats.pct(ship.hullExplRes)}</td>
-            <td className='ri'>{formats.pct(ship.hullKinRes)}</td>
-            <td className='ri'>{formats.pct(ship.hullThermRes)}</td>
-          </tr>
-        </tbody>
-      </table>
-      <br />
-      { ship.shield ?
-        <table style={{ width: '100%', lineHeight: '1em', backgroundColor: 'transparent' }}>
-          <thead>
-            <tr>
-              <th colSpan={3}>{translate('shields')}</th>
-            </tr>
-            <tr>
-              <th>{translate('strength')}</th>
-              <th onMouseEnter={termtip.bind(null, 'PHRASE_SG_RECOVER', { cap: 0 })} onMouseLeave={hide}>{translate('recovery')}</th>
-              <th onMouseEnter={termtip.bind(null, 'PHRASE_SG_RECHARGE', { cap: 0 })} onMouseLeave={hide}>{translate('recovery')}</th>
-            </tr>
-          </thead>
+      <span>
+        <h1>{translate('defence summary')}</h1>
+        <table className='summary' style={{ width: '100%', lineHeight: '1em', backgroundColor: 'transparent' }}>
           <tbody>
+            {ship.shield ?
             <tr>
-              <td className='ri'>{formats.int(ship.shield)} {units.MJ}</td>
-              <td className='ri'>{ship.shield ? formats.time(ship.calcShieldRecovery()) : '-'}</td>
-              <td className='ri'>{ship.shield ? formats.time(ship.calcShieldRecharge()) : '-'}</td>
+              <td colSpan='4' className='summary'><h2>{translate('shields')}: {formats.int(ship.shield)} {units.MJ}</h2></td>
+            </tr> : null }
+            {ship.shield ?
+            <tr>
+              <td className='le'>{translate('damage from')}</td>
+              <td className='ri' onMouseOver={termtip.bind(null, 'explosive')} onMouseOut={tooltip.bind(null, null)}><DamageExplosive /> {formats.pct1(ship.shieldExplRes || 1)}</td>
+              <td className='ri' onMouseOver={termtip.bind(null, 'kinetic')} onMouseOut={tooltip.bind(null, null)}><DamageKinetic /> {formats.pct1(ship.shieldKinRes || 1)}</td>
+              <td className='ri' onMouseOver={termtip.bind(null, 'thermal')} onMouseOut={tooltip.bind(null, null)}><DamageThermal /> {formats.pct1(ship.shieldThermRes || 1)}</td>
+            </tr> : null }
+
+            { ship.shield && ship.shieldCells ?
+            <tr>
+              <td colSpan='4'><h2>{translate('shield cells')}: {formats.int(ship.shieldCells)} {units.MJ}</h2></td>
+            </tr> : null }
+
+            <tr>
+              <td colSpan='4'><h2>{translate('armour')}: {formats.int(ship.armour)}</h2></td>
+            </tr>
+
+            <tr>
+              <td className='le'>{translate('damage from')}</td>
+              <td className='ri' onMouseOver={termtip.bind(null, 'explosive')} onMouseOut={tooltip.bind(null, null)}><DamageExplosive /> {formats.pct1(ship.hullExplRes || 1)}</td>
+              <td className='ri' onMouseOver={termtip.bind(null, 'kinetic')} onMouseOut={tooltip.bind(null, null)}><DamageKinetic /> {formats.pct1(ship.hullKinRes || 1)}</td>
+              <td className='ri' onMouseOver={termtip.bind(null, 'thermal')} onMouseOut={tooltip.bind(null, null)}><DamageThermal /> {formats.pct1(ship.hullThermRes || 1)}</td>
             </tr>
           </tbody>
-        </table> : null }
-      </div>
+        </table>
+      </span>
     );
   }
 }