+++ src/app/i18n/en.js
@@ -92,14 +92,18 @@ export const terms = {
   // Unit for seconds
   secs: 's',
 
-  // Hardpoint abbreviations
+  // Weapon, offence and defence
   dpe: 'Damage per MJ of energy',
   dps: 'Damage per second',
+  sdps: 'Sustained damage per second',
   dpssdps: 'Damage per second (sustained damage per second)',
   eps: 'Energy per second',
   epsseps: 'Energy per second (sustained energy per second)',
   hps: 'Heat per second',
   hpsshps: 'Heat per second (sustained heat per second)',
+  'damage by': 'Damage by',
+  'damage from': 'Damage from',
+  'shield cells': 'Shield cells',
 
   // Modifications
   ammo: 'Ammunition maximum',
@@ -134,6 +138,7 @@ export const terms = {
   rof: 'Rate of fire',
   shield: 'Shield',
   shieldboost: 'Shield boost',
+  shieldreinforcement: 'Shield reinforcement',
   spinup: 'Spin up time',
   syscap: 'Systems capacity',
   sysrate: 'Systems recharge rate',